requirements.txt
asgiref==3.8.1
certifi==2024.8.30
charset-normalizer==3.3.2
Django==5.1.1
django-intl-tel-input==0.4.0
django-phonenumber-field==8.0.0
django-widget-tweaks==1.5.0
djangorestframework==3.15.2
idna==3.10
phonenumbers==8.13.46
requests==2.32.3
sqlparse==0.5.1
pip~=24.0
filelock~=3.13.1
Jinja2~=3.1.3
urllib3~=2.1.0
django-tracking-jl~=0.5.8
python-dateutil~=2.9.0
